+++ app/Http/Controllers/Ticket.php
@@ -35,6 +35,10 @@ class Ticket extends Controller
      */
     public function store(Request $request)
     {
+        if(!Session::get('Login') || Session::get('LoginRole') != 'Master' || Session::get('LoginRole') != 'EventOrganizer')
+        {
+            return redirect('/login/master')->with('status', 'You have to login first!');
+        }
         $request->validate([
             'name' => 'required|max:64',
             'amount' => 'required',
@@ -74,6 +78,11 @@ class Ticket extends Controller
      */
     public function edit($id)
     {
+        if(!Session::get('Login') || Session::get('LoginRole') != 'Master')
+        {
+            return redirect('/login/master')->with('status', 'You have to login first!');
+        }
+
         $ticket = TModel::find($id);
         $colors = [
             'primary',
@@ -96,6 +105,11 @@ class Ticket extends Controller
      */
     public function update(Request $request, $id)
     {
+        if(!Session::get('Login') || Session::get('LoginRole') != 'Master' || Session::get('LoginRole') != 'EventOrganizer')
+        {
+            return redirect('/login/master')->with('status', 'You have to login first!');
+        }
+
         $request->validate([
             'name' => 'required|max:64',
             'amount' => 'required',
@@ -123,6 +137,10 @@ class Ticket extends Controller
      */
     public function destroy($id)
     {
+        if(!Session::get('Login') || Session::get('LoginRole') != 'Master')
+        {
+            return redirect('/login/master')->with('status', 'You have to login first!');
+        }
         $ticket = TModel::find($id);
         TModel::destroy($id);
